Make Button Container block a functional component
